+++ wiringPi/wiringShift.h
@@ -35,7 +35,7 @@ extern "C" {
 
 extern uint8_t shiftIn  (uint8_t dPin, uint8_t cPin, uint8_t order) ;
 extern void    shiftOut (uint8_t dPin, uint8_t cPin, uint8_t order, uint8_t val) ;
-extern void	   shiftOutWithDelay (uint8_t dPin, uint8_t cPin, uint8_t order, uint8_t val, uint8_t delay) ;
+extern void	   shiftOutWithDelay (uint8_t dPin, uint8_t cPin, uint8_t order, uint8_t val, int delay) ;
 
 #ifdef __cplusplus
 }
